+++ porte_monnaie_citoyen_server/routes/event.js
@@ -8,11 +8,19 @@ const generatePassword = require('generate-password');
  * Récupère tous les "events"
  */
 router.get("/", (req, res) => {
-    _Event.find()
-    .populate("town")
-    .populate("partner")
-    .then(events => res.json(events))
-    .catch(error => res.json({error : "Impossible de récupérer les events."}))
+    let user = req._passport.session.user;
+    if(req.isAuthenticated() && (user.role === 'town' || user.role === 'citizen')){
+        _Event.find()
+            .populate("town")
+            .populate("partner")
+            .then(events => res.json(events))
+            .catch(error => res.json({error : "Impossible de récupérer les events."}))
+    }
+    else{
+        res.status(401);
+        res.json({error : "Vous n'avez pas les autorisations pour effectuer cette action"});
+    }
+    
 });
 
 /**
@@ -20,11 +28,18 @@ router.get("/", (req, res) => {
  * Récupère un "event" par rapport à son id
  */
 router.get("/:id", (req, res) => {
-    _Event.findOne({_id : req.params.id})
-    .populate("town")
-    .populate("partner")
-    .then(event => res.json(event))
-    .catch(error => res.json({error : "Impossible de récupérer l'event"}))
+    let role = req._passport.session.user;
+    if(req.isAuthenticated() && (user.role === 'town' || user.role === 'citizen')){
+        _Event.findOne({_id : req.params.id})
+            .populate("town")
+            .populate("partner")
+            .then(event => res.json(event))
+            .catch(error => res.json({error : "Impossible de récupérer l'event"}))
+    }
+    else{
+        res.status(401);
+        res.json({error : "Vous n'avez pas les autorisations pour effectuer cette action"});
+    }    
 });
 
 /**
@@ -32,22 +47,29 @@ router.get("/:id", (req, res) => {
  * Créé un nouvel "event"
  */
 router.post("/", (req, res) => {
-    let newEvent = new _Event({
-        name: req.body.name,
-        description: req.body.description,
-        address: req.body.address,
-        gain: 0,
-        password: generatePassword.generate({length: 24, numbers: true}),
-        date: new Date(req.body.date),
-        image: "placeholder.png",
-        state: "W",
-        town: req.body.town,
-        partner: req.body.partner
-    });
-
-    newEvent.save()
-    .then(event => res.json(event))
-    .catch(error => res.json({error}))
+    let user = req._passport.session.user;
+    if(req.isAuthenticated() && user.role === 'partner' && req.body.partner._id === user.id){
+        let newEvent = new _Event({
+            name: req.body.name,
+            description: req.body.description,
+            address: req.body.address,
+            gain: 0,
+            password: generatePassword.generate({length: 24, numbers: true}),
+            date: new Date(req.body.date),
+            image: "placeholder.png",
+            state: "W",
+            town: req.body.town,
+            partner: req.body.partner
+        });
+        newEvent.save()
+        .then(event => res.json(event))
+        .catch(error => res.json({error}))
+    }
+    else{
+        res.status(401);
+        res.json({error : "Vous n'avez pas les autorisations pour effectuer cette action"});
+    }
+    
 });
 
 /**
@@ -55,9 +77,16 @@ router.post("/", (req, res) => {
  * Modifie un "event"
  */
 router.put("/", (req, res) => {
-    _Event.updateOne({_id : req.body.event._id}, req.body.event)
-    .then(event => res.json(req.body.event._id))
-    .catch(error => res.json({error : "Impossible de mettre à jour l'event"}))
+    let user = req._passport.session.user;
+    if(req.isAuthenticated() && user.role === 'partner'){
+        _Event.updateOne({_id : req.body.event._id}, req.body.event)
+        .then(event => res.json(req.body.event._id))
+        .catch(error => res.json({error : "Impossible de mettre à jour l'event"}))
+    }
+    else{
+        res.status(401);
+        res.json({error : "Vous n'avez pas les autorisations pour effectuer cette action"});
+    }   
 });
 
 /**
@@ -65,9 +94,17 @@ router.put("/", (req, res) => {
  * Supprime un "event"
  */
 router.delete("/:id", (req, res) => {
-    _Event.deleteOne({_id : req.params.id})
-    .then(result => res.json(result))
-    .catch(error => res.json({error : "Impossible de supprimer l'event"}))
+    let user = req._passport.session.user;
+    if(req.isAuthenticated() && user.role === 'partner'){
+        _Event.deleteOne({_id : req.params.id})
+            .then(result => res.json(result))
+            .catch(error => res.json({error : "Impossible de supprimer l'event"}))
+    }
+    else{
+        res.status(401);
+        res.json({error : "Vous n'avez pas les autorisations pour effectuer cette action"});
+    }
+    
 });
 
 module.exports = router;
